WLM changes complete
git-svn-id: https://svn.forschooner.net/code/dnx/trunk/dnx@110 2be5450d-c085-40c6-83eb-9dc9c13f2dba

+++ common/dnxLogging.c
@@ -35,7 +35,7 @@
 
 #define MAX_LOG_LINE 1023
 
-static long defDebug = 0;
+static int defDebug = 0;
 static int defLogFacility = LOG_LOCAL7;
 
 static int * pDebug = &defDebug;
